api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.1
mysql-connector-python==8.4.0
